added the system and converted doubles to floats

+++ components/GravityWell.h
@@ -1,9 +1,10 @@
 #ifndef COMPONENTS_GRAVITY_WELL_H
 #define COMPONENTS_GRAVITY_WELL_H
 
-struct GravityWell {
-    double mass;
-    double r_s; // schwartzschild radius
+struct GravityWell
+{
+    float mass;
+    float r_s; // schwartzschild radius
 };
 
 #endif
