+++ components/Projectile.h
@@ -1,9 +1,9 @@
 #ifndef COMPONENTS_PROJECTILE_H
 #define COMPONENTS_PROJECTILE_H
 
-
-struct Projectile {
-    double impactParameter;
+struct Projectile
+{
+    float impactParameter;
 };
 
 #endif
