+++ deprecated-code/BlackHole.h
@@ -5,8 +5,8 @@
 
 class Blackhole : Object {
     public:
-        double r_s; // schwartzschild radius
-        double mass;
-        Blackhole(vec2 pos, double m) : mass(m), Object(pos) {r_s = 2 * G * mass /(C * C);};
+        float r_s; // schwartzschild radius
+        float mass;
+        Blackhole(vec2 pos, float m) : mass(m), Object(pos) {r_s = 2 * G * mass /(C * C);};
         void draw() override;
 };
